separarted timetable and course routes into the new structure
