+++ server/src/timetable/timetable.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { TimetableController } from './timetable.controller';
+import { TimetableService } from './timetable.service';
+
+@Module({
+  controllers: [TimetableController],
+  providers: [TimetableService],
+})
+export class TimetableModule {}
