+++ server/src/user/types.ts
@@ -15,20 +15,3 @@ export class UserSettings {
   unscheduleClassesByDefault: boolean;
   hideExamClasses: boolean;
 }
-
-export class AddCourseDto {
-  term: string;
-  colour: string;
-}
-export class CourseDetails {
-  id: string;
-  selectedClasses: string[];
-}
-
-export class UserTimetable {
-  id: string;
-  name: string;
-  year: number;
-  term: string;
-  primary: boolean;
-}
